example/requirements.txt
Pillow
django==2.2.16
django-composite-field==1.1.0
pytest
pytest-django
pytest-factoryboy
pytest-cov
pytest-isort
pytest-black
